pyproject.toml
[project]
name = "pygame-chess"
version = "0.1.0"
description = "python chess game using pygame"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.5",
]
